requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
SQLAlchemy==2.0.23
alembic==1.13.0
paypalrestsdk==1.13.1
python-dotenv==1.0.0
psycopg2-binary==2.9.10
colorlog==6.8.0
gunicorn==22.0.0
greenlet==3.0.0
